Implement a --verbose flag
Normally, kdumpid will finish as soon as it can find out the kernel
architecture and version. However, if this is not sufficient, it will
search the data and also print out additional fields.

The --verbose flag can be used to display the additional data even if
the header was sufficient to find out the kernel arch and version.