Fix handling of CUSTOM_CFLAGS
CUSTOM_CFLAGS may be changed in the Makefile or given on the command
line. They should be combined with the default CFLAGS (which define
the version, for example).

This was screwed up in the Makefile, giving the user no option to
override the flags (without breaking the flags used internally).