Use -lz unconditionally
The file util.c uses inflate* functions, so there's no point in testing
for libz: we need it regardless. If the test in libs.sh fails, then -lz
will not be appended to the link, and linking will fail. So remove the
test and add -lz to the Makefile regardless.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>